Fixed overflow bug when explaining stupidly slow queries

+++ intermine/src/java/org/intermine/sql/query/ExplainResult.java
@@ -144,7 +144,7 @@ public class ExplainResult
      */
     public long getTime() {
         // TODO: Do this properly.
-        return (complete * estimatedRows) / rows;
+        return (long) ((((double) complete) * ((double) estimatedRows)) / ((double) rows));
     }
 
     /**
